summaryclipper_api/requirements.txt
Flask==2.2.3
openai==0.27.2
pytube==15.0.0
pydub==0.25.1
openai-whisper==20230314
python-dotenv==1.0.0